uacpi shims tweaks

+++ include/uacpi_arch_helpers.h
@@ -2,13 +2,14 @@
 #define UACPI_ARCH_HELPERS_H
 
 #include <uacpi/platform/atomic.h>
+#include <uacpi_types.h>
 
 #ifndef UACPI_ARCH_FLUSH_CPU_CACHE
 #define UACPI_ARCH_FLUSH_CPU_CACHE() do {} while (0)
 #endif
 
-typedef unsigned long uacpi_cpu_flags;
-typedef unsigned long long uacpi_thread_id;
+typedef uint8_t uacpi_cpu_flags;
+typedef uint64_t uacpi_thread_id;
 
 #ifndef UACPI_ATOMIC_LOAD_THREAD_ID
 #define UACPI_ATOMIC_LOAD_THREAD_ID(ptr) ((uacpi_thread_id)uacpi_atomic_load64(ptr))
